# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kamwaalay API",
        "description": "API documentation for Kamwaalay - A platform connecting households with trusted domestic helpers",
        "contact": {
            "email": "contact@kamwaalay.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard",
                "description": "Get admin dashboard statistics and recent activity. Requires admin role.",
                "operationId": "2c29ca3389addbd18db1fff0f925cfd3",
                "responses": {
                    "200": {
                        "description": "Admin dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stats": {
                                            "description": "Platform statistics",
                                            "type": "object"
                                        },
                                        "recent_bookings": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "pending_helpers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/helpers": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List helpers (admin)",
                "description": "Get paginated list of helpers with optional status filter. Requires admin role.",
                "operationId": "20ad3bc48e8d189197a20d2c595c7232",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "verified",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of helpers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "helpers": {
                                            "description": "Paginated list of helpers",
                                            "type": "object"
                                        },
                                        "filters": {
                                            "description": "Applied filters",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/helpers/{helper}/status": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update helper status",
                "description": "Update helper verification status and active status. Requires admin role.",
                "operationId": "99a8b70bced9430d427b2b567bc42f48",
                "parameters": [
                    {
                        "name": "helper",
                        "in": "path",
                        "description": "Helper user ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "verification_status"
                                ],
                                "properties": {
                                    "verification_status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "verified",
                                            "rejected"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Helper status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Helper status updated successfully!"
                                        },
                                        "helper": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Helper not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/bookings": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List bookings (admin)",
                "description": "Get paginated list of all bookings with optional status filter. Requires admin role.",
                "operationId": "34c684e10699731bca2840d71f096b71",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "in_progress",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "bookings": {
                                            "description": "Paginated list of bookings",
                                            "type": "object"
                                        },
                                        "filters": {
                                            "description": "Applied filters",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/documents": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List documents (admin)",
                "description": "Get paginated list of all documents with optional status filter. Requires admin role.",
                "operationId": "0c8ec189c8473583340f3a576a70050e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "verified",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of documents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "documents": {
                                            "description": "Paginated list of documents",
                                            "type": "object"
                                        },
                                        "filters": {
                                            "description": "Applied filters",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/documents/{document}/status": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update document status",
                "description": "Update document verification status. Requires admin role. If all documents are verified, helper is automatically verified.",
                "operationId": "b3e1da59361c46b9021972b4d019951b",
                "parameters": [
                    {
                        "name": "document",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "verified",
                                            "rejected"
                                        ]
                                    },
                                    "admin_notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document status updated successfully!"
                                        },
                                        "document": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    },
...
}